+++ handler/Gopkg.lock
@@ -15,7 +15,7 @@
   version = "v1.13.2"
 
 [[projects]]
-  digest = "1:1befaee8931b3f14b9b3d6eee29c1cd4601def8981869c281ebb98dd8d6dc5d9"
+  digest = "1:32d6bbcac8b563ddd5ddee033186313d81ecdac1dbc49d7ea08a3530f012655c"
   name = "github.com/aws/aws-sdk-go"
   packages = [
     "aws",
@@ -80,21 +80,20 @@
     "service/kms/kmsiface",
     "service/organizations",
     "service/rds",
-    "service/rds/rdsiface",
     "service/s3",
     "service/s3/s3iface",
     "service/s3/s3manager",
     "service/secretsmanager",
-    "service/secretsmanager/secretsmanageriface",
     "service/sns",
+    "service/sns/snsiface",
     "service/sqs",
     "service/ssm",
     "service/sts",
     "service/sts/stsiface",
   ]
   pruneopts = "UT"
-  revision = "23e996a57757e6b586048809309e3945557c4edb"
-  version = "v1.25.22"
+  revision = "3550bcc6d4824489d3463ba8642bad0c1f4e3755"
+  version = "v1.25.25"
 
 [[projects]]
   digest = "1:7b94d37d65c0445053c6f3e73090e3966c1c29127035492c349e14f25c440359"
@@ -196,12 +195,12 @@
   version = "v1.4.0"
 
 [[projects]]
-  digest = "1:7f60bb033c03f8322869de36b3a063080593d22037c5a117007553441645adaa"
+  digest = "1:777b07148bdabfc846ab120fd0d19b88820cb56578d24a776ec5bf42bafb7ac3"
   name = "github.com/tealeg/xlsx"
   packages = ["."]
   pruneopts = "UT"
-  revision = "1b556911422e5b0f499a31ddc028ccc5ad288d4c"
-  version = "v1.0.5"
+  revision = "626b0bf6d25fddacbb58ae127b37e4de04268f84"
+  version = "v2.0.0"
 
 [[projects]]
   branch = "master"
@@ -226,7 +225,7 @@
   name = "golang.org/x/net"
   packages = ["context"]
   pruneopts = "UT"
-  revision = "fe3aa8a4527195a6057b3fad46619d7d090e99b5"
+  revision = "0deb6923b6d97481cb43bc1043fe5b72a0143032"
 
 [[projects]]
   branch = "master"
@@ -234,7 +233,7 @@
   name = "golang.org/x/sys"
   packages = ["cpu"]
   pruneopts = "UT"
-  revision = "f43be2a4598cf3a47be9f94f0c28197ed9eae611"
+  revision = "e8c54fb511f6a2d695a4403ae0dab1c05aef011a"
 
 [[projects]]
   digest = "1:c25289f43ac4a68d88b02245742347c94f1e108c534dda442188015ff80669b3"
@@ -282,13 +281,12 @@
     "github.com/aws/aws-sdk-go/service/kms/kmsiface",
     "github.com/aws/aws-sdk-go/service/organizations",
     "github.com/aws/aws-sdk-go/service/rds",
-    "github.com/aws/aws-sdk-go/service/rds/rdsiface",
     "github.com/aws/aws-sdk-go/service/s3",
     "github.com/aws/aws-sdk-go/service/s3/s3iface",
     "github.com/aws/aws-sdk-go/service/s3/s3manager",
     "github.com/aws/aws-sdk-go/service/secretsmanager",
-    "github.com/aws/aws-sdk-go/service/secretsmanager/secretsmanageriface",
     "github.com/aws/aws-sdk-go/service/sns",
+    "github.com/aws/aws-sdk-go/service/sns/snsiface",
     "github.com/aws/aws-sdk-go/service/ssm",
     "github.com/aws/aws-sdk-go/service/sts",
     "github.com/aws/aws-sdk-go/service/sts/stsiface",
