+++ handler/Gopkg.toml
@@ -31,7 +31,7 @@
 
 [[constraint]]
   name = "github.com/aws/aws-sdk-go"
-  version = "1.25.22"
+  version = "1.25.25"
 
 [[constraint]]
   name = "github.com/caarlos0/env"
@@ -43,7 +43,7 @@
 
 [[constraint]]
   name = "github.com/tealeg/xlsx"
-  version = "1.0.5"
+  version = "2.0.0"
 
 [prune]
   go-tests = true
