+++ handler/Makefile
@@ -37,7 +37,7 @@ test: lint
 	go test -v ./...
 
 test_nocreds: lint
-	go test -v --run 'Users|Groups|Policies|Roles|Vaults|TestInstances|Images|Volumes|TestSnapshots|Vpcs|Subnets|Addresses|KeyPairs|Buckets|Stacks|Alarms|ConfigRules|LoadBalancers|Keys' ./...
+	go test -v --run 'Users|Groups|Policies|Roles|Vaults|TestInstances|Images|Volumes|TestSnapshots|Vpcs|Subnets|Addresses|KeyPairs|Buckets|Stacks|Alarms|ConfigRules|LoadBalancers|Keys|Subscriptions|Topics' ./...
 
 lint: Gopkg.toml dependencies
 	dep ensure
